Remove zentral module dependency on vpc module
Used because google_service_networking_connection is required. This
resource has been moved from the vpc module to the zentral module, and
the dependencies can be better declared now.

This probably resolves an issue with the zentral module
google_client_config data source, which was not behaving correctly
because it was depending on other resources.

WARNING: you need to manually move the 2 resources in a current state
with terraform state mv.