Use autohealing health checks for web & worker MIG
Better control of the instance lifecycle using HTTP autohealing health
checks for the web and worker MIG. For the web instances, we target the
/instance-health-check nginx endpoint, that starts answering with 200s
when the ztl_admin setup is done. For the worker instances, we target
the first worker prometheus endpoint (/metrics on port 9910), that is
started toward the end of the setup.